cast doctests to plain python objects

+++ speclite/filters.py
@@ -564,8 +564,8 @@ class FilterResponse(object):
     returning a numpy array of response values:
 
     >>> resp = rband([5980, 6000, 6020])
-    >>> np.float64(np.round(resp[1], 4))
-    np.float64(0.4692)
+    >>> float(np.round(resp[1], 4))
+    0.4692
 
     The effective wavelength of a filter is defined as the
     :ref:`photon-weighted <weights>` mean wavelength:
@@ -653,8 +653,6 @@ class FilterResponse(object):
         Linear interpolator of our response function that returns zero for
         all values outside our wavelength range.  Should normally be evaluated
         through our :meth:`__call__` convenience method.
-
-
 
     Raises
     ------
