+++ speclite/redshift.py
@@ -35,8 +35,8 @@ def redshift(z_in, z_out, data_in=None, data_out=None, rules=[]):
     ... dict(name='flux', exponent=-1, array_in=flux)])
     >>> result.dtype
     dtype([('wlen', '<f8'), ('flux', '<f8')])
-    >>> np.float64(result['flux'][0])
-    np.float64(0.5)
+    >>> float(result['flux'][0])
+    0.5
 
     The same calculation could be performed with the input data stored in
     a numpy structured array, in which case any additional fields are
@@ -51,8 +51,8 @@ def redshift(z_in, z_out, data_in=None, data_out=None, rules=[]):
     ... dict(name='flux', exponent=-1)])
     >>> result.dtype
     dtype([('wlen', '<f8'), ('flux', '<f8'), ('maskbits', '<i8')])
-    >>> result['flux'][0]
-    np.float64(0.5)
+    >>> float(result['flux'][0])
+    0.5
 
     The transformed result is always a `numpy structured array
     <http://docs.scipy.org/doc/numpy/user/basics.rec.html>`__, with field
